Move mdx-plus-react to basic
