+++ packages/integrations/mdx/test/fixtures/basic/astro.config.mjs
@@ -1,7 +1,8 @@
 import { defineConfig } from 'astro/config';
+import react from "@astrojs/react";
 import mdx from "@astrojs/mdx";
 
 export default defineConfig({
   site: 'https://mdx-is-neat.com/',
-  integrations: [mdx()]
+  integrations: [react(), mdx()]
 });
