+++ packages/integrations/mdx/test/fixtures/basic/src/components/plus-react/Component.jsx
