+++ packages/integrations/mdx/test/fixtures/basic/src/pages/plus-react/index.astro
@@ -1,5 +1,5 @@
 ---
-import Component from "../components/Component.jsx";
+import Component from "../../components/plus-react/Component.jsx";
 ---
 <html>
 	<head>
