--- packages/integrations/mdx/test/fixtures/mdx-plus-react/astro.config.mjs
@@ -1,6 +1,0 @@
-import mdx from '@astrojs/mdx';
-import react from '@astrojs/react';
-
-export default {
-	integrations: [react(), mdx()]
-}
