--- packages/integrations/mdx/test/fixtures/mdx-plus-react/package.json
@@ -1,11 +1,0 @@
-{
-  "name": "@test/mdx-plus-react",
-  "private": true,
-  "dependencies": {
-    "@astrojs/mdx": "workspace:*",
-    "@astrojs/react": "workspace:*",
-    "astro": "workspace:*",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0"
-  }
-}
