--- packages/integrations/mdx/test/mdx-plus-react.test.js
@@ -1,25 +1,0 @@
-import mdx from '@astrojs/mdx';
-
-import { expect } from 'chai';
-import { parseHTML } from 'linkedom';
-import { loadFixture } from '../../../astro/test/test-utils.js';
-
-describe('MDX and React', () => {
-	let fixture;
-
-	before(async () => {
-		fixture = await loadFixture({
-			root: new URL('./fixtures/mdx-plus-react/', import.meta.url),
-		});
-		await fixture.build();
-	});
-
-	it('can be used in the same project', async () => {
-		const html = await fixture.readFile('/index.html');
-		const { document } = parseHTML(html);
-
-		const p = document.querySelector('p');
-
-		expect(p.textContent).to.equal('Hello world');
-	});
-});
